[Feat]#66-컨테이너 뷰 분리 & 연결 확인
