+++ Dokgi/View/AddVerseView/AddVerseVC.swift
@@ -20,12 +20,16 @@ protocol BookSelectionDelegate: AnyObject {
 class AddVerseVC: UIViewController {
     
     let viewModel = AddVerseViewModel()
+    let containerView = AddVerseContainerLayout()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        keywordCollectionView.register(KeywordCell.self, forCellWithReuseIdentifier: KeywordCell.reuseIdentifier)
-        keywordField.delegate = self
+        containerView.keywordCollectionView.register(KeywordCell.self, forCellWithReuseIdentifier: KeywordCell.reuseIdentifier)
+        containerView.keywordCollectionView.delegate = self // 데이터소스와 델리게이트 설정
+        containerView.keywordCollectionView.dataSource = self // 데이터소스와 델리게이트 설정
+        containerView.keywordField.delegate = self
+        containerView.verseTextView.delegate = self
         setupViews()
         initLayout()
         setupActions()
@@ -41,328 +45,35 @@ class AddVerseVC: UIViewController {
     
     let scrollView = UIScrollView().then {
         $0.showsVerticalScrollIndicator = false
-    }
-    
-    let viewInScroll = UIView()
-    
-    let scanButton = UIButton(configuration: .filled(), primaryAction: nil).then {
-        $0.configurationUpdateHandler = { button in
-            var configuration = button.configuration
-            configuration?.title = "구절 스캔"
-            configuration?.titleTextAttributesTransformer = UIConfigurationTextAttributesTransformer { incoming in
-                var outgoing = incoming
-                outgoing.font = Pretendard.bold.dynamicFont(style: .subheadline)
-                return outgoing
-            }
-            configuration?.baseForegroundColor = .charcoalBlue
-            configuration?.baseBackgroundColor = .lightSkyBlue
-            configuration?.image = UIImage(resource: .camera).withTintColor(UIColor(resource: .charcoalBlue), renderingMode: .alwaysOriginal)
-            configuration?.imagePadding = 10
-            button.configuration = configuration
-        }
-        $0.layer.cornerRadius = 17
-        $0.clipsToBounds = true
-    }
-    
-    let infoView = UIView().then {
-        $0.layer.cornerRadius = 15
-    }
-    
-    let overlayView = UIView().then {
-        $0.backgroundColor = UIColor.lightSkyBlue.withAlphaComponent(0.5)
-        $0.layer.cornerRadius = 15
-    }
-    
-    let searchButton = UIButton(type: .system).then {
-        var config = UIButton.Configuration.filled()
-        config.title = "책 검색"
-        $0.titleLabel?.font = Pretendard.semibold.dynamicFont(style: .headline)
-        config.baseForegroundColor = .white
-        config.baseBackgroundColor = .charcoalBlue
-        config.image = .magnifyingglass
-        config.imagePadding = 8
-        config.imagePlacement = .leading
-        config.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 8, bottom: 0, trailing: 15)
-        $0.configuration = config
-        $0.layer.cornerRadius = 15
-        $0.clipsToBounds = true
-        $0.titleLabel?.numberOfLines = 1
-    }
-    
-    var imageView = UIImageView().then {
-        $0.image = UIImage(resource: .empty).withRenderingMode(.alwaysTemplate)
-        $0.tintColor = .buttonLightGray
-        $0.layer.cornerRadius = 15
-        $0.clipsToBounds = true
-        $0.contentMode = .scaleAspectFit
-    }
-    
-    var titleLabel = UILabel().then {
-        $0.text = "책 제목"
-        $0.font = Pretendard.bold.dynamicFont(style: .body)
-        $0.textColor = .bookTextGray
-        $0.numberOfLines = 2
-    }
-    
-    var authorLabel = UILabel().then {
-        $0.text = "저자"
-        $0.font = Pretendard.bold.dynamicFont(style: .body)
-        $0.textColor = .bookTextGray
-        $0.numberOfLines = 2
-    }
-    
-    lazy var verseTextView = UITextView().then {
-        $0.text = "텍스트를 입력하세요"
-        $0.layer.borderWidth = 1.0
-        $0.layer.borderColor = UIColor.lightGray.withAlphaComponent(0.7).cgColor
-        $0.textContainerInset = UIEdgeInsets(top: 16.0, left: 16.0, bottom: 16.0, right: 16.0)
-        $0.font = Pretendard.regular.dynamicFont(style: .callout)
-        $0.textColor = .placeholderText
-        $0.layer.cornerRadius = 8
-        $0.clipsToBounds = true
-        $0.showsVerticalScrollIndicator = false
-        $0.delegate = self
-    }
-    
-    let characterCountLabel = UILabel().then {
-        $0.textColor = .gray
-        $0.font = Pretendard.bold.dynamicFont(style: .footnote)
-    }
-    
-    private let pencilImageView = UIImageView().then {
-        $0.image = UIImage(named: "pencil")
-        $0.contentMode = .scaleAspectFit
-    }
-    
-    let keywordLabel = UILabel().then {
-        $0.attributedText = AddVerseVC.createAttributedString(for: "키워드 (선택)")
-        $0.textAlignment = .left
-    }
-    
-    let keywordField = UITextField().then {
-        let placeholder = "키워드를 입력해 주세요"
-        $0.placeholder = placeholder
-        $0.clipsToBounds = true
-        $0.layer.cornerRadius = 10
-        $0.layer.borderWidth = 1
-        $0.layer.borderColor = UIColor.borderGray.cgColor
-    }
-
-    lazy var keywordCollectionView = UICollectionView(frame: .zero, collectionViewLayout: {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        layout.minimumInteritemSpacing = 10
-        layout.minimumLineSpacing = 10
-        return layout
-    }()).then {
-        $0.backgroundColor = .clear
-        $0.delegate = self
-        $0.dataSource = self
-        $0.showsHorizontalScrollIndicator = false
-    }
-    
-    let pageLabel = UILabel().then {
-        $0.text = "페이지"
-        $0.font = Pretendard.semibold.dynamicFont(style: .headline)
-        $0.textColor = .black
-    }
-    
-    let pageNumberTextField = UITextField().then {
-        $0.placeholder = "페이지 수"
-        $0.font = Pretendard.regular.dynamicFont(style: .subheadline)
-        $0.borderStyle = .roundedRect
-    }
-    
-    let betterSegmentedControl: BetterSegmentedControl = {
-        let segmentedControl = BetterSegmentedControl(
-            frame: .zero,
-            segments: LabelSegment.segments(
-                withTitles: ["Page", "%"],
-                normalFont: Pretendard.semibold.dynamicFont(style: .footnote),
-                normalTextColor: .charcoalBlue,
-                selectedFont: Pretendard.semibold.dynamicFont(style: .footnote),
-                selectedTextColor: .white
-            ),
-            options: [
-                .indicatorViewBackgroundColor(.charcoalBlue),
-                .cornerRadius(15),
-                .backgroundColor(.white)
-            ]
-        )
-
-        // 초기 세그먼트 인덱스 설정
-        segmentedControl.setIndex(0, animated: false, shouldSendValueChangedEvent: false)
-        
-        return segmentedControl
-    }()
-
-    let ControlBoder = UIView().then {
-        $0.backgroundColor = .white
-        $0.layer.cornerRadius = 15
-        $0.layer.borderWidth = 1
-        $0.layer.borderColor = UIColor.charcoalBlue.cgColor
-    }
-    
-    let recordButton = UIButton().then {
-        $0.setTitle("기록 하기", for: .normal)
-        $0.titleLabel?.font = Pretendard.bold.dynamicFont(style: .headline)
-        $0.setTitleColor(.white, for: .normal)
-        $0.backgroundColor = .charcoalBlue
-        $0.layer.cornerRadius = 8
+        $0.alwaysBounceVertical = true
+        $0.contentInsetAdjustmentBehavior = .never
     }
     
     func setupViews() {
         view.addSubview(scrollView)
-        scrollView.addSubview(viewInScroll)
-        viewInScroll.addSubview(scanButton)
-        viewInScroll.addSubview(infoView)
-        infoView.addSubview(imageView)
-        infoView.addSubview(titleLabel)
-        infoView.addSubview(authorLabel)
-        infoView.addSubview(overlayView)
-        overlayView.addSubview(searchButton)
-        viewInScroll.addSubview(verseTextView)
-        viewInScroll.addSubview(characterCountLabel)
-        viewInScroll.addSubview(pencilImageView)
-        viewInScroll.addSubview(keywordLabel)
-        viewInScroll.addSubview(keywordField)
-        viewInScroll.addSubview(keywordCollectionView)
-        viewInScroll.addSubview(pageLabel)
-        viewInScroll.addSubview(pageNumberTextField)
-        viewInScroll.addSubview(ControlBoder)
-        viewInScroll.addSubview(betterSegmentedControl)
-        viewInScroll.addSubview(recordButton)
+        scrollView.addSubview(containerView)
     }
     
     // MARK: - 제약조건
     func initLayout() {
         scrollView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide.snp.top)
-            $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom)
-            $0.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
-            $0.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
+            $0.edges.equalTo(view.safeAreaLayoutGuide)
         }
         
-        viewInScroll.snp.makeConstraints {
-            $0.edges.equalTo(scrollView.snp.edges)
+        containerView.snp.makeConstraints {
+            $0.edges.equalTo(scrollView.contentLayoutGuide)
             $0.width.equalTo(scrollView.snp.width)
-            $0.height.equalTo(1000) // 임시로 1000으로 설정
+            $0.height.greaterThanOrEqualTo(view.safeAreaLayoutGuide.snp.height) // 이 부분을 추가
+            $0.height.greaterThanOrEqualTo(1000)
         }
-        
-        scanButton.snp.makeConstraints {
-            $0.top.equalTo(viewInScroll.snp.top).offset(16)
-            $0.trailing.equalTo(viewInScroll.snp.trailing).offset(-16)
-        }
-        
-        infoView.snp.makeConstraints {
-            $0.centerY.equalTo(viewInScroll.snp.top).offset(170)
-            $0.horizontalEdges.equalTo(viewInScroll).inset(16)
-            $0.height.equalTo(200)
-        }
-        
-        overlayView.snp.makeConstraints {
-            $0.edges.equalTo(infoView)
-        }
-        
-        searchButton.snp.makeConstraints {
-            $0.center.equalToSuperview()
-            $0.height.equalTo(35)
-        }
-        
-        imageView.snp.makeConstraints {
-            $0.leading.equalTo(infoView.snp.leading)
-            $0.centerY.equalTo(infoView.snp.centerY)
-            $0.width.equalTo(103)
-            $0.height.equalTo(146)
-        }
-        
-        titleLabel.snp.makeConstraints {
-            $0.leading.equalTo(imageView.snp.trailing).offset(16)
-            $0.centerY.equalTo(infoView.snp.centerY).offset(-16)
-            $0.trailing.equalTo(infoView.snp.trailing).offset(-16)
-        }
-        
-        authorLabel.snp.makeConstraints {
-            $0.leading.equalTo(imageView.snp.trailing).offset(16)
-            $0.top.equalTo(titleLabel.snp.bottom).offset(16)
-            $0.trailing.equalTo(infoView.snp.trailing).offset(-16)
-        }
-        
-        verseTextView.snp.makeConstraints {
-            $0.top.equalTo(infoView.snp.bottom).offset(32)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.height.equalTo(329)
-        }
-        
-        characterCountLabel.snp.makeConstraints {
-            $0.trailing.equalTo(verseTextView.snp.trailing).offset(-16)
-            $0.bottom.equalTo(verseTextView.snp.bottom).offset(-16)
-        }
-        
-        pencilImageView.snp.makeConstraints {
-            $0.bottom.equalTo(verseTextView.snp.bottom).offset(-8)
-            $0.leading.equalTo(verseTextView.snp.leading).offset(8)
-        }
-        
-        keywordLabel.snp.makeConstraints {
-            $0.top.equalTo(verseTextView.snp.bottom).offset(32)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-        }
-        
-        keywordField.snp.makeConstraints {
-            $0.top.equalTo(keywordLabel.snp.bottom).offset(16)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.height.equalTo(36)
-        }
-        
-        keywordCollectionView.snp.makeConstraints {
-            $0.top.equalTo(keywordField.snp.bottom).offset(16)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.height.equalTo(35)
-        }
-        
-        pageLabel.snp.makeConstraints {
-            $0.top.equalTo(keywordCollectionView.snp.bottom).offset(50)
-            $0.leading.equalTo(viewInScroll.snp.leading).offset(16)
-        }
-        
-        pageNumberTextField.snp.makeConstraints {
-            $0.centerY.equalTo(pageLabel.snp.centerY)
-            $0.leading.equalTo(pageLabel.snp.trailing).offset(8)
-            $0.width.equalTo(55)
-            $0.height.equalTo(30)
-        }
-        
-        betterSegmentedControl.snp.makeConstraints {
-            $0.centerY.equalTo(pageLabel.snp.centerY)
-            $0.trailing.equalTo(viewInScroll.snp.trailing).offset(-16)
-            $0.height.equalTo(30)
-            $0.width.equalTo(120)
-        }
-        
-        ControlBoder.snp.makeConstraints {
-            $0.centerY.equalTo(pageLabel.snp.centerY)
-            $0.trailing.equalTo(viewInScroll.snp.trailing).offset(-15)
-            $0.height.equalTo(32)
-            $0.width.equalTo(122)
-        }
-        
-        recordButton.snp.makeConstraints {
-            $0.top.equalTo(pageLabel.snp.bottom).offset(60)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.centerX.equalToSuperview()
-            $0.height.equalTo(50)
-        }
-        
-        scrollView.contentSize = viewInScroll.bounds.size
     }
     
     func setupActions() {
-        scanButton.addTarget(self, action: #selector(scanButtonTapped(_:)), for: .touchUpInside)
-        searchButton.addTarget(self, action: #selector(searchButtonTapped(_:)), for: .touchUpInside)
-        recordButton.addTarget(self, action: #selector(recordButtonTapped(_:)), for: .touchUpInside)
-        betterSegmentedControl.addTarget(self, action: #selector(segmentedControlValueChanged(_:)), for: .valueChanged)
-        keywordField.addTarget(self, action: #selector(textFieldDidChange(_:)), for: .editingChanged)
+        containerView.scanButton.addTarget(self, action: #selector(scanButtonTapped(_:)), for: .touchUpInside)
+        containerView.searchButton.addTarget(self, action: #selector(searchButtonTapped(_:)), for: .touchUpInside)
+        containerView.recordButton.addTarget(self, action: #selector(recordButtonTapped(_:)), for: .touchUpInside)
+        containerView.betterSegmentedControl.addTarget(self, action: #selector(segmentedControlValueChanged(_:)), for: .valueChanged)
+        containerView.keywordField.addTarget(self, action: #selector(textFieldDidChange(_:)), for: .editingChanged)
     }
     
     @objc func scanButtonTapped(_ sender: UIButton) {
@@ -389,24 +100,24 @@ class AddVerseVC: UIViewController {
     
     @objc func recordButtonTapped(_ sender: UIButton) {
         print("기록하기 버튼이 눌렸습니다.")
-        if searchButton.isHidden == false {
+        if containerView.searchButton.isHidden == false {
             viewModel.showAlert(presenter: self, title: "책 정보 기록", message: "책 검색을 눌러 책 정보를 기록해주세요")
             return
         }
         
-        if verseTextView.text.isEmpty || verseTextView.text == "텍스트를 입력하세요" {
+        if containerView.verseTextView.text.isEmpty || containerView.verseTextView.text == "텍스트를 입력하세요" {
             viewModel.showAlert(presenter: self, title: "구절 입력", message: "구절을 입력해 주세요")
             return
         }
         
-        if pageNumberTextField.text?.isEmpty == true {
+        if containerView.pageNumberTextField.text?.isEmpty == true {
             viewModel.showAlert(presenter: self, title: "페이지", message: "페이지를 입력해 주세요")
             return
         }
         
         viewModel.saveVerse(selectedBook: viewModel.selectedBook,
-                            verseText: verseTextView.text ?? "",
-                            pageNumberText: pageNumberTextField.text ?? "",
+                            verseText: containerView.verseTextView.text ?? "",
+                            pageNumberText: containerView.pageNumberTextField.text ?? "",
                             pageType: viewModel.pageType,
                             keywords: viewModel.keywords) { success in
             if success {
@@ -415,8 +126,6 @@ class AddVerseVC: UIViewController {
                 self.viewModel.showAlert(presenter: self, title: "경고", message: "모든 필수 정보를 입력해주세요.")
             }
         }
-        
-
     }
     
     func showAlert(title: String, message: String, completion: (() -> Void)? = nil) {
@@ -428,34 +137,34 @@ class AddVerseVC: UIViewController {
     }
     
     func setUserInfoTextField() {
-        keywordField.backgroundColor = .white
+        containerView.keywordField.backgroundColor = .white
         
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 10, height: keywordField.frame.height))
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 10, height: containerView.keywordField.frame.height))
         
-        keywordField.leftView = paddingView
-        keywordField.leftViewMode = .always
+        containerView.keywordField.leftView = paddingView
+        containerView.keywordField.leftViewMode = .always
         
-        keywordField.rightView = paddingView
-        keywordField.rightViewMode = .always
+        containerView.keywordField.rightView = paddingView
+        containerView.keywordField.rightViewMode = .always
     }
     
     // 초기 텍스트뷰 글자 수 설정
     func updateCharacterCountLabel() {
-        let currentCount = verseTextView.text == "텍스트를 입력하세요" ? 0 : verseTextView.text.count
-        characterCountLabel.text = "\(currentCount)/200"
+        let currentCount = containerView.verseTextView.text == "텍스트를 입력하세요" ? 0 : containerView.verseTextView.text.count
+        containerView.characterCountLabel.text = "\(currentCount)/200"
     }
     
     // 텍스트 속성을 설정하는 함수
-    private static func createAttributedString(for text: String) -> NSAttributedString {
+    static func createAttributedString(for text: String) -> NSAttributedString {
         let attributedString = NSMutableAttributedString(string: text)
         
         // "키워드" 부분에 대한 속성 설정
         let keywordRange = (text as NSString).range(of: "키워드")
-        attributedString.addAttribute(.font, value: Pretendard.semibold.dynamicFont(style: .callout), range: keywordRange)
+        attributedString.addAttribute(.font, value: Pretendard.semibold.dynamicFont(style: .headline), range: keywordRange)
         
         // "선택" 부분에 대한 속성 설정
         let selectionRange = (text as NSString).range(of: "(선택)")
-        attributedString.addAttribute(.font, value: Pretendard.regular.dynamicFont(style: .callout), range: selectionRange)
+        attributedString.addAttribute(.font, value: Pretendard.regular.dynamicFont(style: .headline), range: selectionRange)
         attributedString.addAttribute(.foregroundColor, value: UIColor.gray, range: selectionRange)
         
         return attributedString
@@ -463,17 +172,17 @@ class AddVerseVC: UIViewController {
     
     func displayBookInfo() {
         if let book = viewModel.selectedBook {
-            titleLabel.text = book.title
-            titleLabel.font = Pretendard.semibold.dynamicFont(style: .headline)
-            titleLabel.textColor = .black
-            authorLabel.text = book.author
+            containerView.titleLabel.text = book.title
+            containerView.titleLabel.font = Pretendard.semibold.dynamicFont(style: .headline)
+            containerView.titleLabel.textColor = .black
+            containerView.authorLabel.text = book.author
             if let url = URL(string: book.image) {
-                imageView.kf.setImage(with: url)
-                imageView.contentMode = .scaleAspectFill
+                containerView.imageView.kf.setImage(with: url)
+                containerView.imageView.contentMode = .scaleAspectFill
             }
         }
-        overlayView.isHidden = true
-        searchButton.isHidden = true
+        containerView.overlayView.isHidden = true
+        containerView.searchButton.isHidden = true
     }
     
     func recognizeText(from image: UIImage) {
@@ -490,7 +199,7 @@ class AddVerseVC: UIViewController {
             
             let recognizedStrings = observations.compactMap { $0.topCandidates(1).first?.string }
             DispatchQueue.main.async {
-                self?.verseTextView.text = recognizedStrings.joined(separator: "\n")
+                self?.containerView.verseTextView.text = recognizedStrings.joined(separator: "\n")
             }
         }
         let revision3 = VNRecognizeTextRequestRevision3
@@ -507,7 +216,7 @@ class AddVerseVC: UIViewController {
     }
     func removeKeyword(at indexPath: IndexPath) {
         viewModel.keywords.remove(at: indexPath.item)
-        keywordCollectionView.deleteItems(at: [indexPath])
+        containerView.keywordCollectionView.deleteItems(at: [indexPath])
     }
 }
 
@@ -516,7 +225,7 @@ extension AddVerseVC: UITextFieldDelegate {
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         if let keyword = textField.text, !keyword.isEmpty {
             viewModel.keywords[viewModel.keywords.count - 1] = keyword
-            keywordCollectionView.reloadData()
+            containerView.keywordCollectionView.reloadData()
         }
         textField.text = ""
         textField.resignFirstResponder()
@@ -526,7 +235,7 @@ extension AddVerseVC: UITextFieldDelegate {
     func textFieldDidBeginEditing(_ textField: UITextField) {
         if let text = textField.text, text.isEmpty {
             viewModel.keywords.append("")
-            keywordCollectionView.reloadData()
+            containerView.keywordCollectionView.reloadData()
         }
     }
     
@@ -537,7 +246,7 @@ extension AddVerseVC: UITextFieldDelegate {
             } else {
                 viewModel.keywords[viewModel.keywords.count - 1] = text
             }
-            keywordCollectionView.reloadData()
+            containerView.keywordCollectionView.reloadData()
         }
     }
 }
@@ -555,7 +264,8 @@ extension AddVerseVC: UICollectionViewDelegate, UICollectionViewDataSource, UICo
         }
         
         let reversedIndex = viewModel.keywords.count - 1 - indexPath.item
-        cell.configure(with: viewModel.keywords[reversedIndex])
+        let keyword = viewModel.keywords[reversedIndex]
+        cell.configure(with: keyword)
         cell.backgroundColor = .lightSkyBlue
         
         return cell
@@ -580,6 +290,7 @@ extension AddVerseVC: UITextViewDelegate {
         guard textView.textColor == .placeholderText else { return }
         textView.textColor = .label
         textView.text = nil
+        updateCharacterCountLabel() // 글자 수 업데이트
     }
     
     func textViewDidEndEditing(_ textView: UITextView) {
@@ -592,14 +303,14 @@ extension AddVerseVC: UITextViewDelegate {
     
     func textViewDidChange(_ textView: UITextView) {
         let currentCount = textView.text.count
-        characterCountLabel.text = "\(currentCount)/200"
+        containerView.characterCountLabel.text = "\(currentCount)/200"
         
         // 최대 200자 제한
         if currentCount > 200 {
             textView.text = String(textView.text.prefix(200))
-            characterCountLabel.text = "200/200"
+            containerView.characterCountLabel.text = "200/200"
         } else {
-            characterCountLabel.text = "\(currentCount)/200"
+            containerView.characterCountLabel.text = "\(currentCount)/200"
         }
     }
 }
